random data generator created, time meeasuring fixed with clock_t

+++ main.c
@@ -29,7 +29,7 @@ int main(int argc, char** args) {
 
     printf("Total profit: %i\n", sum_profit);
     printf("Total weight: %i\n", sum_weight);
-    printf("Time taken: %lf seconds\n", (double) bin_problem.time_taken);
+    printf("Time taken: %lf seconds\n", bin_problem.time_taken);
     printf("Items amount: %i\n", bin_problem.items_amount);
 
     free_bin_problem_struct(&bin_problem);
